+++ src/main/java/com/project/trackfit/trainer/dto/PersonalTrainerDTO.java
@@ -1,4 +1,4 @@
-package com.project.trackfit.trainer;
+package com.project.trackfit.trainer.dto;
 
 import com.project.trackfit.subscription.Subscription;
 import jakarta.validation.constraints.NotNull;
