+++ src/main/java/com/project/trackfit/trainer/dto/RetrieveTrainerRequest.java
@@ -1,4 +1,4 @@
-package com.project.trackfit.trainer;
+package com.project.trackfit.trainer.dto;
 
 import java.util.UUID;
