+++ src/test/java/com/project/trackfit/user/service/ApplicationUserServiceTests.java
@@ -3,7 +3,7 @@ package com.project.trackfit.user.service;
 import com.project.trackfit.user.component.Role;
 import com.project.trackfit.core.exception.EmailAlreadyTakenException;
 import com.project.trackfit.customer.service.ICustomerService;
-import com.project.trackfit.trainer.IPersonalTrainerService;
+import com.project.trackfit.trainer.service.IPersonalTrainerService;
 import com.project.trackfit.user.component.PasswordCreation;
 import com.project.trackfit.user.dto.ApplicationUser;
 import com.project.trackfit.user.dto.CreateUserRequest;
